changed server files to js to avoid issues
